Fix tests and context
